+++ release_notes.md
@@ -3,6 +3,7 @@
 - My change description (#PR)
 -->
 - Return 500 when when writing secrets to storage fails (#8361)
+- Enable diagnostic events collection
 
 **Release sprint:** Sprint 119
 [ [bugs](https://github.com/Azure/azure-functions-host/issues?q=is%3Aissue+milestone%3A%22Functions+Sprint+119%22+label%3Abug+is%3Aclosed) | [features](https://github.com/Azure/azure-functions-host/issues?q=is%3Aissue+milestone%3A%22Functions+Sprint+119%22+label%3Afeature+is%3Aclosed) ]